requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-registration==3.2
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
whitenoise==6.0.0
